t/debug: fix 'fio_debug' declaration
debug.h:38:22: warning: type of ‘fio_debug’ does not match original declaration
 extern unsigned long fio_debug;
                      ^
t/debug.c:5:14: note: previously declared here
 unsigned int fio_debug = 0;

+++ t/debug.c
@@ -2,7 +2,7 @@
 
 FILE *f_err;
 struct timeval *fio_tv = NULL;
-unsigned int fio_debug = 0;
+unsigned long fio_debug = 0;
 
 void __dprint(int type, const char *str, ...)
 {
